006_sixth-hw/task1/src/main/java/org/bagirov/MyIterator.java
package org.bagirov;

import java.util.Iterator;
import java.util.NoSuchElementException;

public class MyIterator<T> implements Iterator<T> {

    private T[] array;
    private int currentIndex = 0;

    public MyIterator(T[] array) {
        this.array = array;
    }

    @Override
    public boolean hasNext() {
        return currentIndex < array.length;
    }

    @Override
    public T next() {
        if (!hasNext()) {
            throw new NoSuchElementException("No element in array");
        }
        return array[currentIndex++];
    }
}
